refactor: 단축키 설명 다이얼로그 기능 비활성화 unused
