+++ components/hospital/icu/header/icu-header.tsx
@@ -1,5 +1,4 @@
 import IcuHeaderDateSelector from '@/components/hospital/icu/header/date-picker/header-date-selector'
-import { InstructionDialog } from '@/components/hospital/icu/header/instruction/instruction-dialog'
 import type { Vet } from '@/types/icu/chart'
 import RegisterDialog from './register-dialog/register-dialog'
 
@@ -14,7 +13,8 @@ export default async function IcuHeader({
 }) {
   return (
     <div className="fixed top-2 z-30 flex w-full items-center justify-center gap-2 px-2 xl:ml-12 xl:w-auto xl:justify-start 2xl:top-1.5 2xl:ml-0 2xl:w-auto">
-      <InstructionDialog />
+      {/* 단축키 안내 다이얼로그 삭제 */}
+      {/* <InstructionDialog /> */}
 
       <RegisterDialog hosId={hosId} vetsData={vetsData} groupList={groupList} />
       <IcuHeaderDateSelector />
