+++ components/hospital/icu/header/instruction(unused)/(unused)/instruction-contents.tsx
