+++ components/hospital/icu/header/instruction(unused)/(unused)/instruction-sidebar.tsx
