+++ components/hospital/icu/header/instruction(unused)/instruction-shortcut-item.tsx
